% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mean_posterior_pred}
\alias{plot_mean_posterior_pred}
\title{plot empirical cohort-based or cumulative mean vs posterior mean}
\usage{
plot_mean_posterior_pred(summarised_mean, obs_mean, obs_at, alpha = 0.3, ...)
}
\arguments{
\item{summarised_mean}{Summarised mean as produced by \code{\link[=summarise_variable]{summarise_variable()}}}

\item{...}{Additional arguments passed to \code{\link[ggplot2:aes]{ggplot2::aes()}}.}

\item{data}{data used for object fitting}

\item{truncate}{account for truncation?}
}
\description{
plot empirical cohort-based or cumulative mean vs posterior mean
}
